Add support for creation and completion dates
This fixes https://github.com/JMoerman/Go-For-It/issues/17.

The format specification states that a completion date is mandatory and a
creation date can be optionally added. The format states that "Many Todo.txt
clients discard priority on task completion." This is one of them partly in
order to simplify the date parsing logic but also it's the logical thing to do.